+++ contracts/src/interfaces/IEpochable.sol
@@ -21,7 +21,7 @@ interface IEpochable {
 
     function interval() external view returns (uint256 res);
 
-    function currentEpochId() external view returns (uint64 res);
+    function currentEpochId() external view returns (uint48 res);
 
     function epochFromBlocknum(uint256 blocknum) external view returns (EpochMath.Epoch memory res);
 }
